refactor: split ec2_helpers.rs into ec2_tags and ec2_security_group_rules modules
Split the 725-line ec2_helpers.rs into two focused modules:
- ec2_tags.rs: EC2 tag helper functions (ec2_tags_to_value, value_to_ec2_tags, apply_ec2_tags)
- ec2_security_group_rules.rs: shared security group rule CRUD operations and convert_protocol_value

Part of #777.
